made drawTerrain not iterate over all positions twice when path is being previewed. Instead it blits the previewed path onto a seperate surface and blits that over the map at the end.

+++ src/Battlescape/Position.h
@@ -90,7 +90,7 @@ public:
 	}
 
 	/// Calculates the distance in 3d.
-	static float distance(Position pos1, Position pos2)
+	static double distance(Position pos1, Position pos2)
 	{
 		return std::sqrt(distanceSq(pos1, pos2));
 	}
